requirements.txt
streamlit
pandas
PyPDF2
python-docx
groq
scikit-learn
